code refactor,add security compile options

+++ src/mpl2mpl/include/muid_replacement.h
@@ -106,7 +106,7 @@ class MUIDReplacement : public FuncOptimizeImpl {
   void GenerateGlobalRootList();
   void CollectImplicitUndefClassInfo(StmtNode &stmt);
   void ReplaceMethodMetaFuncAddr(MIRSymbol &funcSymbol, int64 index);
-  void ReplaceFieldMetaStaticAddr(MIRSymbol &mirSymbol, int64 index);
+  void ReplaceFieldMetaStaticAddr(MIRSymbol &mirSymbol, uint32 index);
   void CollectFuncAndDataFromKlasses();
   void CollectFuncAndDataFromGlobalTab();
   void CollectFuncAndDataFromFuncList();
